Refactor : Service 하나로 합치기
